[chore] rename story
